Resolved errors in Runway.java and RunwayHandling.java

+++ src/main/java/sim/model/stores/Runway.java
@@ -10,12 +10,12 @@ public class Runway {
     private String occupied;
     private RunwayMode mode;
     private RunwayStatus status;
-    private int timeRemaining;
+    private double timeRemaining;
 
     public enum RunwayMode { NONE, LANDING, TAKEOFF, MIXED }
     public enum RunwayStatus { NONE, AVAILABLE, INSPECTION, SNOW, FAILURE }
 
-    public Runway(int _ID,RunwayMode _mode,RunwayStatus _status,int _timeRemaining){
+    public Runway(int _ID,RunwayMode _mode,RunwayStatus _status,double _timeRemaining){
         ID = _ID;
         occupied = "";
         mode = _mode;
